+++ uCNC/src/hal/mcus/esp32/mcu_esp32.c
@@ -24,7 +24,6 @@
 #include "esp_ipc.h"
 #include "driver/uart.h"
 #include "driver/timer.h"
-#include "soc/i2s_struct.h"
 #include <string.h>
 #include <stdbool.h>
 #include <stdint.h>
@@ -33,232 +32,13 @@
 
 static volatile bool esp32_global_isr_enabled;
 static volatile bool mcu_itp_timer_running;
-#ifdef IC74HC595_CUSTOM_SHIFT_IO
-volatile uint32_t ic74hc595_i2s_pins;
-#endif
+
 hw_timer_t *esp32_step_timer;
 
 MCU_CALLBACK void mcu_itp_isr(void *arg);
-static MCU_CALLBACK void mcu_gen_pwm_and_servo(void);
-static MCU_CALLBACK void mcu_gen_step(void);
+MCU_CALLBACK void mcu_gen_pwm_and_servo(void);
+MCU_CALLBACK void mcu_gen_step(void);
 MCU_CALLBACK void mcu_gpio_isr(void *type);
-
-/**
- * IO 74HC595 expander via I2S
- * **/
-#ifdef IC74HC595_CUSTOM_SHIFT_IO
-#if IC74HC595_COUNT != 4
-#error "IC74HC595_COUNT must be 4(bytes) to use ESP32 I2S mode for IO shifting"
-#endif
-#include "driver/i2s.h"
-#include "soc/i2s_struct.h"
-
-#ifndef I2S_SAMPLE_RATE
-#define I2S_SAMPLE_RATE (F_STEP_MAX * 2)
-#endif
-#define I2S_SAMPLES_PER_BUFFER (I2S_SAMPLE_RATE / 500) // number of samples per 2ms (0.002/1 = 1/500)
-#define I2S_BUFFER_COUNT 5														 // DMA buffer size 5 * 2ms = 10ms stored motions (can be adjusted but may cause to much or too little latency)
-#define I2S_SAMPLE_US (1000000UL / I2S_SAMPLE_RATE)		 // (1s/250KHz = 0.000004s = 4us)
-
-#ifdef ITP_SAMPLE_RATE
-#undef ITP_SAMPLE_RATE
-#endif
-#define ITP_SAMPLE_RATE (I2S_SAMPLE_RATE)
-
-static volatile uint32_t i2s_mode;
-#define I2S_MODE __atomic_load_n((uint32_t *)&i2s_mode, __ATOMIC_RELAXED)
-
-// software generated oneshot for RT steps like laser PPI
-#if defined(MCU_HAS_ONESHOT_TIMER) && defined(ENABLE_RT_SYNC_MOTIONS)
-static uint32_t esp32_oneshot_counter;
-static uint32_t esp32_oneshot_reload;
-static FORCEINLINE void mcu_gen_oneshot(void)
-{
-	if (esp32_oneshot_counter)
-	{
-		esp32_oneshot_counter--;
-		if (!esp32_oneshot_counter)
-		{
-			if (mcu_timeout_cb)
-			{
-				mcu_timeout_cb();
-			}
-		}
-	}
-}
-#endif
-
-// implements the custom step mode function to switch between buffered stepping and realtime stepping
-uint8_t itp_set_step_mode(uint8_t mode)
-{
-	uint8_t last_mode = I2S_MODE;
-	if (mode)
-	{
-		itp_sync();
-#ifdef USE_I2S_REALTIME_MODE_ONLY
-		__atomic_store_n((uint32_t *)&i2s_mode, (ITP_STEP_MODE_SYNC | ITP_STEP_MODE_REALTIME), __ATOMIC_RELAXED);
-#else
-		__atomic_store_n((uint32_t *)&i2s_mode, (ITP_STEP_MODE_SYNC | mode), __ATOMIC_RELAXED);
-#endif
-		cnc_delay_ms(20);
-	}
-	return last_mode;
-}
-
-static void IRAM_ATTR esp32_i2s_stream_task(void *param)
-{
-	int8_t available_buffers = I2S_BUFFER_COUNT;
-	i2s_event_t evt;
-	portTickType xLastWakeTimeUpload = xTaskGetTickCount();
-	i2s_config_t i2s_config = {
-			.mode = I2S_MODE_MASTER | I2S_MODE_TX, // Only TX
-			.sample_rate = I2S_SAMPLE_RATE,
-			.bits_per_sample = I2S_BITS_PER_SAMPLE_32BIT,
-			.channel_format = I2S_CHANNEL_FMT_ONLY_LEFT, // 1-channels
-			.communication_format = I2S_COMM_FORMAT_STAND_I2S | I2S_COMM_FORMAT_STAND_MSB,
-			.dma_buf_count = I2S_BUFFER_COUNT,
-			.dma_buf_len = I2S_SAMPLES_PER_BUFFER,
-			.use_apll = false,
-			.intr_alloc_flags = ESP_INTR_FLAG_LEVEL1, // Interrupt level 1
-			.tx_desc_auto_clear = false,
-			.fixed_mclk = 0};
-
-	i2s_pin_config_t pin_config = {
-			.bck_io_num = IC74HC595_I2S_CLK,
-			.ws_io_num = IC74HC595_I2S_WS,
-			.data_out_num = IC74HC595_I2S_DATA,
-			.data_in_num = -1 // Not used
-	};
-	QueueHandle_t i2s_dma_queue;
-
-	i2s_driver_install(IC74HC595_I2S_PORT, &i2s_config, I2S_BUFFER_COUNT, &i2s_dma_queue);
-	i2s_set_pin(IC74HC595_I2S_PORT, &pin_config);
-
-	for (;;)
-	{
-		uint32_t mode = I2S_MODE;
-
-		// tracks DMA buffer usage
-		if (available_buffers < I2S_BUFFER_COUNT)
-		{
-			while (xQueueReceive(i2s_dma_queue, &evt, 0) == pdPASS)
-			{
-				if (evt.type == I2S_EVENT_TX_DONE)
-				{
-					available_buffers++;
-				}
-			}
-		}
-
-		// updates the working mode
-		if (mode & ITP_STEP_MODE_SYNC)
-		{
-			// wait for DMA to output content
-			while (available_buffers < I2S_BUFFER_COUNT)
-			{
-				// tracks DMA buffer usage
-				while (xQueueReceive(i2s_dma_queue, &evt, 0) == pdPASS)
-				{
-					if (evt.type == I2S_EVENT_TX_DONE)
-					{
-						available_buffers++;
-					}
-				}
-				vTaskDelayUntil(&xLastWakeTimeUpload, (20 / portTICK_RATE_MS));
-			}
-
-			switch (mode & ~ITP_STEP_MODE_SYNC)
-			{
-			case ITP_STEP_MODE_DEFAULT:
-				// timer_pause(ITP_TIMER_TG, ITP_TIMER_IDX);
-				// timer_disable_intr(ITP_TIMER_TG, ITP_TIMER_IDX);
-				I2SREG.conf.tx_start = 0;
-				I2SREG.conf.tx_reset = 1;
-				I2SREG.conf.tx_reset = 0;
-				I2SREG.conf.rx_fifo_reset = 1;
-				I2SREG.conf.rx_fifo_reset = 0;
-				available_buffers = I2S_BUFFER_COUNT;
-				i2s_driver_uninstall(IC74HC595_I2S_PORT);
-				i2s_driver_install(IC74HC595_I2S_PORT, &i2s_config, I2S_BUFFER_COUNT, &i2s_dma_queue);
-				i2s_set_pin(IC74HC595_I2S_PORT, &pin_config);
-				break;
-			case ITP_STEP_MODE_REALTIME:
-				I2SREG.conf.tx_start = 0;
-				I2SREG.conf.tx_reset = 1;
-				I2SREG.conf.tx_reset = 0;
-				I2SREG.conf.rx_fifo_reset = 1;
-				I2SREG.conf.rx_fifo_reset = 0;
-				available_buffers = I2S_BUFFER_COUNT;
-				// modify registers for realtime usage
-				I2SREG.out_link.stop = 1;
-				I2SREG.fifo_conf.dscr_en = 0;
-				I2SREG.conf.tx_start = 0;
-				I2SREG.int_clr.val = 0xFFFFFFFF;
-				I2SREG.clkm_conf.clka_en = 0;			 // Use PLL/2 as reference
-				I2SREG.clkm_conf.clkm_div_num = 2; // reset value of 4
-				I2SREG.clkm_conf.clkm_div_a = 1;	 // 0 at reset, what about divide by 0?
-				I2SREG.clkm_conf.clkm_div_b = 0;	 // 0 at reset
-				I2SREG.fifo_conf.tx_fifo_mod = 3;	 // 32 bits single channel data
-				I2SREG.conf_chan.tx_chan_mod = 3;	 //
-				I2SREG.sample_rate_conf.tx_bits_mod = 32;
-				I2SREG.conf.tx_msb_shift = 0;
-				I2SREG.conf.rx_msb_shift = 0;
-				I2SREG.int_ena.out_eof = 0;
-				I2SREG.int_ena.out_dscr_err = 0;
-				I2SREG.conf_single_data = __atomic_load_n((uint32_t *)&ic74hc595_i2s_pins, __ATOMIC_RELAXED);
-				I2SREG.conf1.tx_stop_en = 0;
-				I2SREG.int_ena.val = 0;
-				I2SREG.fifo_conf.dscr_en = 1;
-				I2SREG.int_clr.val = 0xFFFFFFFF;
-				I2SREG.out_link.start = 1;
-				I2SREG.conf.tx_start = 1;
-				ets_delay_us(20);
-				break;
-			}
-
-			// clear sync flag
-			__atomic_fetch_and((uint32_t *)&i2s_mode, ~ITP_STEP_MODE_SYNC, __ATOMIC_RELAXED);
-		}
-
-		while (mode == ITP_STEP_MODE_DEFAULT && available_buffers > 0)
-		{
-			uint32_t i2s_dma_buffer[I2S_SAMPLES_PER_BUFFER];
-
-			for (uint32_t t = 0; t < I2S_SAMPLES_PER_BUFFER; t++)
-			{
-#if defined(IC74HC595_HAS_STEPS) || defined(IC74HC595_HAS_DIRS)
-				mcu_gen_step();
-#endif
-#if defined(IC74HC595_HAS_PWMS) || defined(IC74HC595_HAS_SERVOS)
-				mcu_gen_pwm_and_servo();
-#endif
-#if defined(MCU_HAS_ONESHOT_TIMER) && defined(ENABLE_RT_SYNC_MOTIONS)
-				mcu_gen_oneshot();
-#endif
-				// write to buffer
-				i2s_dma_buffer[t] = __atomic_load_n((uint32_t *)&ic74hc595_i2s_pins, __ATOMIC_RELAXED);
-			}
-
-			uint32_t w = 0;
-
-			i2s_write(IC74HC595_I2S_PORT, &i2s_dma_buffer[0], I2S_SAMPLES_PER_BUFFER * 4, &w, portMAX_DELAY);
-			available_buffers--;
-		}
-
-		vTaskDelayUntil(&xLastWakeTimeUpload, (5 / portTICK_RATE_MS));
-	}
-}
-
-static FORCEINLINE void esp32_i2s_extender_init(void)
-{
-#ifdef USE_I2S_REALTIME_MODE_ONLY
-	itp_set_step_mode(ITP_STEP_MODE_REALTIME);
-#else
-	itp_set_step_mode(ITP_STEP_MODE_DEFAULT);
-#endif
-	xTaskCreatePinnedToCore(esp32_i2s_stream_task, "esp32I2Supdate", 4096, NULL, 7, NULL, CONFIG_ARDUINO_RUNNING_CORE);
-}
-#endif
 
 #ifndef ITP_SAMPLE_RATE
 #define ITP_SAMPLE_RATE (F_STEP_MAX * 2)
@@ -359,7 +139,7 @@ static FORCEINLINE void servo_update(void)
 }
 #endif
 
-static MCU_CALLBACK void mcu_gen_pwm_and_servo(void)
+MCU_CALLBACK void mcu_gen_pwm_and_servo(void)
 {
 	static int16_t mcu_soft_io_counter;
 	int16_t t = mcu_soft_io_counter;
@@ -401,7 +181,7 @@ static MCU_CALLBACK void mcu_gen_pwm_and_servo(void)
 
 static volatile uint32_t mcu_itp_timer_reload;
 static volatile bool mcu_itp_timer_running;
-static MCU_CALLBACK void mcu_gen_step(void)
+MCU_CALLBACK void mcu_gen_step(void)
 {
 	static bool step_reset = true;
 	static int32_t mcu_itp_timer_counter;
@@ -624,7 +404,7 @@ void mcu_init(void)
 	timer_start(ITP_TIMER_TG, ITP_TIMER_IDX);
 
 #ifdef IC74HC595_CUSTOM_SHIFT_IO
-	esp32_i2s_extender_init();
+	mcu_i2s_extender_init();
 #endif
 
 	// initialize rtc timer (currently on core 1)
