+++ uCNC/src/hal/mcus/esp32common/esp32_common.h
@@ -25,6 +25,9 @@ extern "C"
 #endif
 
 #ifdef ESP32
+extern volatile uint32_t i2s_mode;
+#define I2S_MODE __atomic_load_n((uint32_t *)&i2s_mode, __ATOMIC_RELAXED)
+
 void __attribute__((weak)) mcu_uart_init(void){}
 void __attribute__((weak)) mcu_uart_start(void){}
 void __attribute__((weak)) mcu_uart_dotasks(void){}
@@ -42,6 +45,7 @@ void __attribute__((weak)) mcu_wifi_init(void){}
 void __attribute__((weak)) mcu_wifi_dotasks(void){}
 void __attribute__((weak)) mcu_bt_init(void){}
 void __attribute__((weak)) mcu_bt_dotasks(void){}
+void mcu_i2s_extender_init(void);
 #endif
 
 #ifdef __cplusplus
